+++ src/main/java/br/gov/model/cobranca/CobrancaDocumento.java
@@ -28,10 +28,10 @@ public class CobrancaDocumento implements Serializable {
 	private Integer id;
 	
 	@Column(name="dotp_id")
-	private Short documentoTipo;
+	private Integer documentoTipo;
 	
 	@Column(name="cdst_id")
-	private Short cobrancaDebitoSituacao;
+	private Integer cobrancaDebitoSituacao;
 	
 	@Column(name="cbdo_tmemissao")
 	private Date emissao;
@@ -55,19 +55,19 @@ public class CobrancaDocumento implements Serializable {
 		this.id = id;
 	}
 
-	public Short getDocumentoTipo() {
+	public Integer getDocumentoTipo() {
 		return documentoTipo;
 	}
 
-	public void setDocumentoTipo(Short documentoTipo) {
+	public void setDocumentoTipo(Integer documentoTipo) {
 		this.documentoTipo = documentoTipo;
 	}
 
-	public Short getCobrancaDebitoSituacao() {
+	public Integer getCobrancaDebitoSituacao() {
 		return cobrancaDebitoSituacao;
 	}
 
-	public void setCobrancaDebitoSituacao(Short cobrancaDebitoSituacao) {
+	public void setCobrancaDebitoSituacao(Integer cobrancaDebitoSituacao) {
 		this.cobrancaDebitoSituacao = cobrancaDebitoSituacao;
 	}
