+++ src/main/java/br/gov/model/faturamento/Conta.java
@@ -53,7 +53,7 @@ public class Conta implements Serializable{
 	private Short digitoVerificadorConta;
 	
 	@Column(name="dcst_idatual")
-	private Short debitoCreditoSituacaoAtual;
+	private Integer debitoCreditoSituacaoAtual;
 	
 	@Column(name="cnta_nnleituraanterior")
 	private Integer leituraAnterior;
@@ -135,7 +135,7 @@ public class Conta implements Serializable{
 	private BigDecimal percentualEsgoto;
 	
 	@Column(name="cnta_pccoleta")
-	private BigDecimal percentualColeta;
+	private Short percentualColeta;
 	
 	@Column(name="cnta_vlrateioagua")
 	private BigDecimal valorRateioAgua;
@@ -249,7 +249,7 @@ public class Conta implements Serializable{
 
 	public static class Builder{
 		private Integer          consumoAguaEsgoto = 0;
-		private Short            debitoCreditoSituacaoAtual;
+		private Integer          debitoCreditoSituacaoAtual;
 		private Date             dataVencimentoConta;
 		private Date             dataValidadeConta = new Date();
 		private FaturamentoGrupo faturamentoGrupo;
@@ -257,7 +257,7 @@ public class Conta implements Serializable{
 		private Short            indicadorCobrancaMulta = (short) 2;
 		private Short            indicadorAlteracaoVencimento;
 		private BigDecimal       percentualEsgoto;
-		private BigDecimal       percentualColeta;
+		private Short            percentualColeta;
 		private Integer          referencia;
 		private Integer          referenciaContabil;
 		private Rota             rota;
@@ -345,7 +345,7 @@ public class Conta implements Serializable{
 			return this;
 		}
 		
-		public Builder percentualColeta(BigDecimal percentual) {
+		public Builder percentualColeta(Short percentual) {
 			percentualColeta = percentual;
 			return this;
 		}
@@ -531,11 +531,11 @@ public class Conta implements Serializable{
 		this.quadra = quadra;
 	}
 
-	public Short getDebitoCreditoSituacaoAtual() {
+	public Integer getDebitoCreditoSituacaoAtual() {
 		return debitoCreditoSituacaoAtual;
 	}
 
-	public void setDebitoCreditoSituacaoAtual(Short debitoCreditoSituacaoAtual) {
+	public void setDebitoCreditoSituacaoAtual(Integer debitoCreditoSituacaoAtual) {
 		this.debitoCreditoSituacaoAtual = debitoCreditoSituacaoAtual;
 	}
 
@@ -643,11 +643,11 @@ public class Conta implements Serializable{
 		this.percentualEsgoto = percentualEsgoto;
 	}
 
-	public BigDecimal getPercentualColeta() {
+	public Short getPercentualColeta() {
 		return percentualColeta;
 	}
 
-	public void setPercentualColeta(BigDecimal percentualColeta) {
+	public void setPercentualColeta(Short percentualColeta) {
 		this.percentualColeta = percentualColeta;
 	}
